# swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "API esterne per MeatBALLS",
    "version" : "1.0.0",
    "title" : "MeatBALLS API",
    "license" : {
      "name" : "MIT",
      "url" : "https://opensource.org/licenses/MIT"
    }
  },
  "host" : "localhost:5000",
  "basePath" : "/apimeatballs",
  "schemes" : [ "http" ],
  "paths" : {
    "/apimeatballs/getrecipe/{sub_name}" : {
      "get" : {
        "summary" : "Returns a recipe including the parameter 'sub_name'in the meal name as JSON object",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "sub_name",
          "in" : "path",
          "description" : "Sub name of recipe",
          "required" : true,
          "type" : "string",
          "example" : "pizza"
        },{
          "name" : "max_calories",
          "in" : "query",
          "descritpiton" : "Max number of calories",
          "required" : true,
          "type" : "int",
          "example" : "2500"
        },{
          "name" : "excluded_ingredients",
          "in" : "query",
          "descritpiton" : "Ingredeints to exlude, for intollerances, preferences etc..",
          "required" : false,
          "type" : "string, value separated by '-' "
        },{
          "name" : "included_ingredients",
          "in" : "query",
          "descritpiton" : "Ingredients to include",
          "required" : false,
          "type" : "string, value separated by '-' "
        }],
        "responses" : {
          "200" : {
            "description": "No error, success",
            "schema": {
              "type": "json",
              "example":{
                "title":"Quick and Easy St. Louis-Style Pizza",
                "calories": "281",
                "protein":"12g",
                "fat":"12g",
                "carbs":"29g",
                "recipe":"\nStep 1):\nPreheat oven to ..."
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          },
          "500" : {
            "description" : "Error in External API query"
          }
        }
      }
    },
    "/apimeatballs/nutritionalvalues" : {
      "get" : {
        "summary" : "Shows nutritional values of this ingredient as HTML page",
        "produces" : [ "text/html" ],
        "parameters" : [ {
          "name" : "ingredient",
          "in" : "query",
          "description" : "Ingredient to analyze",
          "required" : true,
          "type" : "string",
          "example":"flour"
        } ],
        "responses" : {
          "200" : {
            "description": "No error, success",
            "schema": {
              "type": "string",
              "example":"<div itemprop=\"nutrition\" itemscope itemtype=\"http://schema.org/NutritionInformation..."
            }
          },
          "400" : {
            "description" : "Bad Request"
          },
          "500" : {
            "description" : "Error in External API query"
          }
        }
      }
    },
    "/apimeatballs/shoppinglist/:product_type" : {
      "get" : {
        "summary" : "Returns 10 produtcs which you can buy in a grocery store in an ARRAY of JSON objects",
        "produces" : [ "application/json" ],
        "parameters" : [{
          "name" : "product_type",
          "in": "path",
          "description" : "type of product required",
          "required" : true,
          "type" : "string",
          "example" : "butter"
        }
        ,{
          "name" : "max_calories",
          "in" : "query",
          "description" : "Max amount of calories",
          "required" : true,
          "type" : "int",
          "example" : "2500"
        },
        {
          "name" : "max_fat",
          "in" : "query",
          "description" : "Max amount of fat",
          "required" : true,
          "type" : "int",
          "example" : "200"
          }],
        "responses" : {
          "200" : {
            "description": "No error, success",
            "schema": {
              "type": "json",
              "example": [{
                "title":"Tnuva Butter",
                "image":"https://spoonacular.com/productImages/406748-312x231.jpg",
                "imageType":"jpg"
              },{
                "title":"other product",
                "image":"...",
                "imageType":"..."
              }]
            }
          },
          "400" : {
            "description" : "Bad Request"
          },
          "500" : {
            "description" : "Error in External API query"
          }
        }
      }
    }
  },
  "definitions" : { }
}